+++ internal/integration/stream_test_helpers.go
@@ -497,10 +497,6 @@ func closeConnectors(t testing.TB, env *streamTestEnvironment, input input.Strea
 		output.TriggerCloseNow()
 		require.NoError(t, output.WaitForClose(env.ctx))
 	}
-	closeInput(t, env, input)
-}
-
-func closeInput(t testing.TB, env *streamTestEnvironment, input iinput.Streamed) {
 	if input != nil {
 		input.TriggerStopConsuming()
 		require.NoError(t, input.WaitForClose(env.ctx))